# ARROW-8772: [C++] Unrolled aggregate dense for better speculative execution