1. Expand SumKernel benchmark to more types(Float, Double, Int8, Int16, Int32, Int64).
2. Unrolled the aggregate kernel dense part to speculative add the result in parrel.

Signed-off-by: Frank Du <frank.du@intel.com>