@jianxind 2020-05-25
We find the SumKernel benchmark dense(null percent 0) results is relatively low compared to sparse part for float and double type. Below is the result before unrolled the loop.
SumKernelFloat/32768/0         9.12 us         9.11 us        76809 bytes_per_second=3.35021G/s null_percent=0 size=32.768k
SumKernelFloat/32768/1         4.38 us         4.38 us       159861 bytes_per_second=6.97134G/s null_percent=1 size=32.768k
SumKernelDouble/32768/0        4.99 us         4.99 us       140190 bytes_per_second=6.11937G/s null_percent=0 size=32.768k
SumKernelDouble/32768/1        2.14 us         2.14 us       327897 bytes_per_second=14.246G/s null_percent=1 size=32.768k

With the unroll change,  the dense sumkernel benchmark get 3.7x for float and 2.6x for double.
SumKernelFloat/32768/0         2.46 us         2.46 us       285185 bytes_per_second=12.4269G/s null_percent=0 size=32.768k
SumKernelDouble/32768/0        1.90 us         1.90 us       370921 bytes_per_second=16.0846G/s null_percent=0 size=32.768k

Anyway, it can get more higher performance if using intrinsic, I'd like to work at later point.