@wesm 2020-05-25
@jianxind thank you for looking into this! I had suspected that there was room for improvement in this algorithm. You're certainly free to implement kernels and put them in aggregate_basic_$SIMD_VERSION.cc, and then we can utilize the CpuInfo from `ExecContext` during kernel dispatch to select a supported kernel.